backend/requirements.txt
Django==3.2
django-colorfield==0.8.0
django-environ==0.9.0
django-filter==22.1     
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
gunicorn==20.1.0
Pillow==9.3.0
psycopg2-binary==2.9.6
reportlab==3.6.12
webcolors==1.12
